♻️(backend) extract BBB api calls to a module
Actual BBB API calls are now made from a dedicated module.
This will allow us to use another platform, and decouples it from ours
API views.
