+++ src/backend/marsha/bbb/tests/test_api.py
@@ -4,19 +4,18 @@ import json
 import random
 import re
 from unittest import mock
-from unittest.mock import MagicMock
 from urllib.parse import quote_plus
 import uuid
 
 from django.test import TestCase, override_settings
 
-import requests
-import responses
 from rest_framework_simplejwt.tokens import AccessToken
 
+from marsha.bbb import api, serializers
 from marsha.core import factories as core_factories
 
 from ..factories import MeetingFactory
+from ..utils.bbb_utils import ApiMeetingException
 
 
 # We don't enforce arguments documentation in tests
@@ -31,17 +30,14 @@ class MeetingAPITest(TestCase):
 
     maxDiff = None
 
-    @mock.patch.object(requests, "get")
-    def test_api_meeting_fetch_student(self, mock_create_request):
+    @mock.patch.object(serializers, "get_meeting_infos")
+    def test_api_meeting_fetch_student(self, mock_get_meeting_infos):
         """A student should be allowed to fetch a meeting."""
         meeting = MeetingFactory()
-        mock_create_request.return_value.status_code = 200
-        mock_create_request.return_value.content = """
-        <response>
-            <returncode>SUCCESS</returncode>
-            <running>true</running>
-        </response>
-        """
+        mock_get_meeting_infos.return_value = {
+            "returncode": "SUCCESS",
+            "running": "true",
+        }
 
         jwt_token = AccessToken()
         jwt_token.payload["resource_id"] = str(meeting.id)
@@ -71,17 +67,14 @@ class MeetingAPITest(TestCase):
             content,
         )
 
-    @mock.patch.object(requests, "get")
-    def test_api_meeting_fetch_instructor(self, mock_create_request: MagicMock):
+    @mock.patch.object(serializers, "get_meeting_infos")
+    def test_api_meeting_fetch_instructor(self, mock_get_meeting_infos):
         """An instructor should be able to fetch a meeting."""
         meeting = MeetingFactory()
-        mock_create_request.return_value.status_code = 200
-        mock_create_request.return_value.content = """
-        <response>
-            <returncode>SUCCESS</returncode>
-            <running>true</running>
-        </response>
-        """
+        mock_get_meeting_infos.return_value = {
+            "returncode": "SUCCESS",
+            "running": "true",
+        }
 
         jwt_token = AccessToken()
         jwt_token.payload["resource_id"] = str(meeting.id)
@@ -163,17 +156,15 @@ class MeetingAPITest(TestCase):
         )
         self.assertEqual(response.status_code, 403)
 
-    @mock.patch.object(requests, "get")
-    def test_api_meeting_update_instructor(self, mock_create_request: MagicMock):
+    @mock.patch.object(serializers, "get_meeting_infos")
+    def test_api_meeting_update_instructor(self, mock_get_meeting_infos):
         """An instructor should be able to update a meeting."""
         meeting = MeetingFactory()
-        mock_create_request.return_value.status_code = 200
-        mock_create_request.return_value.content = """
-        <response>
-            <returncode>SUCCESS</returncode>
-            <running>true</running>
-        </response>
-        """
+
+        mock_get_meeting_infos.return_value = {
+            "returncode": "SUCCESS",
+            "running": "true",
+        }
 
         jwt_token = AccessToken()
         jwt_token.payload["resource_id"] = str(meeting.id)
@@ -289,8 +280,8 @@ class MeetingAPITest(TestCase):
             response.json(),
         )
 
-    @mock.patch.object(requests, "get")
-    def test_api_bbb_create_anonymous(self, mock_create_request: MagicMock):
+    @mock.patch.object(api, "create")
+    def test_api_bbb_create_anonymous(self, mock_create_request):
         """An anonymous should not be able to create a meeting."""
         meeting = MeetingFactory()
 
@@ -298,8 +289,8 @@ class MeetingAPITest(TestCase):
         self.assertEqual(response.status_code, 401)
         mock_create_request.assert_not_called()
 
-    @mock.patch.object(requests, "get")
-    def test_api_bbb_create_user_logged_in(self, mock_create_request: MagicMock):
+    @mock.patch.object(api, "create")
+    def test_api_bbb_create_user_logged_in(self, mock_create_request):
         """A logged in user should not be able to create a meeting."""
         user = core_factories.UserFactory(
             first_name="Jane", last_name="Doe", email="jane.doe@example.com"
@@ -311,8 +302,8 @@ class MeetingAPITest(TestCase):
         self.assertEqual(response.status_code, 401)
         mock_create_request.assert_not_called()
 
-    @mock.patch.object(requests, "get")
-    def test_api_bbb_create_student(self, mock_create_request: MagicMock):
+    @mock.patch.object(api, "create")
+    def test_api_bbb_create_student(self, mock_create_request):
         """A student should not be able to create a meeting."""
         meeting = MeetingFactory()
 
@@ -327,35 +318,15 @@ class MeetingAPITest(TestCase):
         self.assertEqual(response.status_code, 403)
         mock_create_request.assert_not_called()
 
-    @mock.patch.object(requests, "get")
-    def test_api_bbb_create_new_meeting(self, mock_create_request: MagicMock):
-        """A new meeting should be started if no BBB meeting exists for the same id."""
-        meeting = MeetingFactory(
-            meeting_id="21e6634f-ab6f-4c77-a665-4229c61b479a",
-            title="Meeting 1",
-            attendee_password="9#R1kuUl3R",
-            moderator_password="0$C7Aaz0o",
-        )
-        mock_create_request.return_value.status_code = 200
-        mock_create_request.return_value.content = f"""
-        <response>
-            <returncode>SUCCESS</returncode>
-            <meetingID>{str(meeting.id)}</meetingID>
-            <internalMeetingID>232a8ab5dbfde4d33a2bd9d5bbc08bd74d04e163-1628693645640</internalMeetingID>
-            <parentMeetingID>bbb-none</parentMeetingID>
-            <attendeePW>9#R1kuUl3R</attendeePW>
-            <moderatorPW>0$C7Aaz0o</moderatorPW>
-            <createTime>1628693645640</createTime>
-            <voiceBridge>83267</voiceBridge>
-            <dialNumber>613-555-1234</dialNumber>
-            <createDate>Wed Aug 11 14:54:05 UTC 2021</createDate>
-            <hasUserJoined>false</hasUserJoined>
-            <duration>0</duration>
-            <hasBeenForciblyEnded>false</hasBeenForciblyEnded>
-            <messageKey></messageKey>
-            <message></message>
-        </response>
-        """
+    @mock.patch.object(api, "create")
+    @mock.patch.object(serializers, "get_meeting_infos")
+    def test_api_bbb_create_new_meeting(
+        self, mock_get_meeting_infos, mock_create_request
+    ):
+        """Starting a meeting with parameters should store them."""
+        meeting = MeetingFactory()
+        mock_get_meeting_infos.return_value = {"returncode": "SUCCESS"}
+        mock_create_request.return_value = {"returncode": "SUCCESS"}
 
         jwt_token = AccessToken()
         jwt_token.payload["resource_id"] = str(meeting.id)
@@ -372,42 +343,20 @@ class MeetingAPITest(TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertDictEqual(
             {
-                "attendeePW": meeting.attendee_password,
-                "createDate": "Wed Aug 11 14:54:05 UTC 2021",
-                "createTime": "1628693645640",
-                "dialNumber": "613-555-1234",
-                "duration": "0",
-                "hasBeenForciblyEnded": "false",
-                "hasUserJoined": "false",
-                "internalMeetingID": "232a8ab5dbfde4d33a2bd9d5bbc08bd74d04e163-1628693645640",
-                "meetingID": str(meeting.id),
-                "message": "Meeting created.",
-                "messageKey": None,
-                "moderatorPW": meeting.moderator_password,
-                "parentMeetingID": "bbb-none",
                 "returncode": "SUCCESS",
-                "voiceBridge": "83267",
             },
             response.data,
         )
         meeting.refresh_from_db()
         self.assertEqual("new title", meeting.title)
         self.assertEqual("Hello", meeting.welcome_text)
-        mock_create_request.assert_called_with(
-            "https://10.7.7.1/bigbluebutton/api/create",
-            params={
-                "attendeePW": "9#R1kuUl3R",
-                "checksum": "785ac3521f7bc9af716660c99517add01c2b3f7e",
-                "meetingID": "21e6634f-ab6f-4c77-a665-4229c61b479a",
-                "moderatorPW": "0$C7Aaz0o",
-                "name": "new title",
-                "welcome": "Hello",
-            },
-            verify=True,
-        )
+        mock_create_request.assert_called_with(meeting=meeting)
 
-    @mock.patch.object(requests, "get")
-    def test_api_bbb_create_existing_meeting(self, mock_create_request: MagicMock):
+    @mock.patch.object(api, "create")
+    @mock.patch.object(serializers, "get_meeting_infos")
+    def test_api_bbb_create_existing_meeting(
+        self, mock_get_meeting_infos, mock_create_request
+    ):
         """No new meeting should be started if a BBB meeting exists for the same id."""
         meeting = MeetingFactory(
             meeting_id="21e6634f-ab6f-4c77-a665-4229c61b479a",
@@ -415,14 +364,11 @@ class MeetingAPITest(TestCase):
             attendee_password="9#R1kuUl3R",
             moderator_password="0$C7Aaz0o",
         )
-        mock_create_request.return_value.status_code = 200
-        mock_create_request.return_value.content = """
-        <response>
-            <returncode>FAILED</returncode>
-            <messageKey>idNotUnique</messageKey>
-            <message>A meeting already exists with that meeting ID. Please use a different meeting ID.</message>  # noqa: E501
-        </response>
-        """
+
+        mock_get_meeting_infos.return_value = {"returncode": "SUCCESS"}
+        mock_create_request.side_effect = ApiMeetingException(
+            {"message": "A meeting already exists with that meeting ID."}
+        )
 
         jwt_token = AccessToken()
         jwt_token.payload["resource_id"] = str(meeting.id)
@@ -439,26 +385,15 @@ class MeetingAPITest(TestCase):
         self.assertEqual(response.status_code, 400)
         self.assertDictEqual(
             response.data,
-            {
-                "message": "A meeting already exists with that meeting ID. "
-                "Please use a different meeting ID."
-            },
+            {"message": "A meeting already exists with that meeting ID."},
         )
-        mock_create_request.assert_called_with(
-            "https://10.7.7.1/bigbluebutton/api/create",
-            params={
-                "name": "Meeting 1",
-                "meetingID": "21e6634f-ab6f-4c77-a665-4229c61b479a",
-                "attendeePW": "9#R1kuUl3R",
-                "moderatorPW": "0$C7Aaz0o",
-                "welcome": "Welcome!",
-                "checksum": "34248ddd5ec5df64e4535141d79a71b314b83ecd",
-            },
-            verify=True,
-        )
+        mock_create_request.assert_called_with(meeting=meeting)
 
-    @mock.patch.object(requests, "get")
-    def test_api_bbb_join_meeting_anonymous(self, mock_create_request: MagicMock):
+    @mock.patch.object(api, "join")
+    @mock.patch.object(serializers, "get_meeting_infos")
+    def test_api_bbb_join_meeting_anonymous(
+        self, mock_get_meeting_infos, mock_join_request
+    ):
         """An anonymous should not be able to join a meeting."""
         meeting = MeetingFactory()
 
@@ -466,10 +401,11 @@ class MeetingAPITest(TestCase):
             f"/api/meetings/{meeting.id}/bbb_join/",
         )
         self.assertEqual(response.status_code, 401)
-        mock_create_request.assert_not_called()
+        mock_get_meeting_infos.assert_not_called()
+        mock_join_request.assert_not_called()
 
-    @mock.patch.object(requests, "get")
-    def test_api_bbb_join_meeting_user_logged_in(self, mock_create_request: MagicMock):
+    @mock.patch.object(api, "join")
+    def test_api_bbb_join_meeting_user_logged_in(self, mock_join_request):
         """A logged in user should not be able to join a meeting."""
         user = core_factories.UserFactory(
             first_name="Jane", last_name="Doe", email="jane.doe@example.com"
@@ -479,7 +415,7 @@ class MeetingAPITest(TestCase):
 
         response = self.client.patch(f"/api/meetings/{meeting.id}/bbb_join/")
         self.assertEqual(response.status_code, 401)
-        mock_create_request.assert_not_called()
+        mock_join_request.assert_not_called()
 
     def test_api_bbb_join_student(self):
         """Joining a meeting as student should return an attendee meeting url."""
@@ -560,8 +496,8 @@ class MeetingAPITest(TestCase):
             response.data,
         )
 
-    @mock.patch.object(requests, "get")
-    def test_api_bbb_end_meeting_anonymous(self, mock_create_request: MagicMock):
+    @mock.patch.object(api, "end")
+    def test_api_bbb_end_meeting_anonymous(self, mock_end_request):
         """An anonymous should not be able to end a meeting."""
         meeting = MeetingFactory()
 
@@ -569,10 +505,10 @@ class MeetingAPITest(TestCase):
             f"/api/meetings/{meeting.id}/bbb_end/",
         )
         self.assertEqual(response.status_code, 401)
-        mock_create_request.assert_not_called()
+        mock_end_request.assert_not_called()
 
-    @mock.patch.object(requests, "get")
-    def test_api_bbb_end_meeting_user_logged_in(self, mock_create_request: MagicMock):
+    @mock.patch.object(api, "end")
+    def test_api_bbb_end_meeting_user_logged_in(self, mock_end_request):
         """A logged in user should not be able to end a meeting."""
         user = core_factories.UserFactory(
             first_name="Jane", last_name="Doe", email="jane.doe@example.com"
@@ -584,10 +520,10 @@ class MeetingAPITest(TestCase):
             f"/api/meetings/{meeting.id}/bbb_end/",
         )
         self.assertEqual(response.status_code, 401)
-        mock_create_request.assert_not_called()
+        mock_end_request.assert_not_called()
 
-    @mock.patch.object(requests, "get")
-    def test_api_bbb_end_meeting_student(self, mock_create_request: MagicMock):
+    @mock.patch.object(api, "end")
+    def test_api_bbb_end_meeting_student(self, mock_end_request):
         """A student should not be able to end a meeting."""
         meeting = MeetingFactory()
 
@@ -600,10 +536,10 @@ class MeetingAPITest(TestCase):
             HTTP_AUTHORIZATION=f"Bearer {jwt_token}",
         )
         self.assertEqual(response.status_code, 403)
-        mock_create_request.assert_not_called()
+        mock_end_request.assert_not_called()
 
-    @responses.activate
-    def test_api_bbb_end_meeting_instructor(self):
+    @mock.patch.object(api, "end")
+    def test_api_bbb_end_meeting_instructor(self, mock_end_request):
         """Ending a meeting as instructor should return a moderator meeting url."""
         meeting = MeetingFactory(
             meeting_id="21e6634f-ab6f-4c77-a665-4229c61b479a",
@@ -611,52 +547,11 @@ class MeetingAPITest(TestCase):
             attendee_password="9#R1kuUl3R",
             moderator_password="0$C7Aaz0o",
         )
-
-        # initial end request
-        responses.add(
-            responses.GET,
-            "https://10.7.7.1/bigbluebutton/api/end",
-            match=[
-                responses.matchers.query_param_matcher(
-                    {
-                        "meetingID": "21e6634f-ab6f-4c77-a665-4229c61b479a",
-                        "password": meeting.moderator_password,
-                        "checksum": "3e7b6970d927f542261f18d6c2c10b5d94bcb55c",
-                    }
-                )
-            ],
-            body="""<response>
-                <returncode>SUCCESS</returncode>
-                <messageKey>sentEndMeetingRequest</messageKey>
-                <message>A request to end the meeting was sent.</message>
-            </response>
-            """,
-            status=200,
-        )
-        # first poll request meeting still exists
-        responses.add(
-            responses.GET,
-            "https://10.7.7.1/bigbluebutton/api/getMeetingInfo",
-            body="""<response>
-                <returncode>SUCCESS</returncode>
-                <meetingName>Super bbb meeting</meetingName>
-                <meetingID>b3fc0805-c9fb-4e62-b12d-d4472986406b</meetingID>
-            </response>
-            """,
-            status=200,
-        )
-        # second poll request meeting does not exist
-        responses.add(
-            responses.GET,
-            "https://10.7.7.1/bigbluebutton/api/getMeetingInfo",
-            body="""<response>
-                <returncode>FAILED</returncode>
-                <messageKey>notFound</messageKey>
-                <message>We could not find a meeting with that meeting ID</message>
-            </response>
-            """,
-            status=200,
-        )
+        mock_end_request.return_value = {
+            "message": "A request to end the meeting was sent.",
+            "messageKey": "sentEndMeetingRequest",
+            "returncode": "SUCCESS",
+        }
 
         jwt_token = AccessToken()
         jwt_token.payload["resource_id"] = str(meeting.id)
