+++ src/backend/marsha/bbb/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Marsha bbb app utils modules."""
