+++ src/backend/marsha/bbb/utils/bbb_utils.py
@@ -8,6 +8,8 @@ from django.conf import settings
 import requests
 import xmltodict
 
+from marsha.bbb.models import Meeting
+
 
 logger = logging.getLogger(__name__)
 
@@ -15,9 +17,16 @@ logger = logging.getLogger(__name__)
 class ApiMeetingException(Exception):
     """Exception used when a meeting api request fails."""
 
-    def __init__(meeting, api_response):
+    def __init__(self, api_response):
         """Extract error message from bbb api response."""
         super().__init__(api_response.get("message"))
+
+
+def _password(meeting: Meeting, moderator=False):
+    """Return password for moderator or attendee."""
+    if moderator:
+        return meeting.moderator_password
+    return meeting.attendee_password
 
 
 def sign_parameters(action, parameters):
@@ -38,7 +47,7 @@ def sign_parameters(action, parameters):
     return parameters
 
 
-def request_bbb_api(action, parameters, prepare=False):
+def request_api(action, parameters, prepare=False):
     """Perform generic get request to BBB API."""
     url = f"{settings.BBB_API_ENDPOINT}/{action}"
     signed_parameters = sign_parameters(action, parameters)
@@ -63,7 +72,7 @@ def request_bbb_api(action, parameters, prepare=False):
     raise ApiMeetingException(api_response)
 
 
-def bbb_api_create(meeting):
+def create(meeting: Meeting):
     """Call BBB API to create a meeting."""
     parameters = {
         "meetingID": str(meeting.meeting_id),
@@ -73,41 +82,47 @@ def bbb_api_create(meeting):
         "welcome": meeting.welcome_text,
     }
 
-    api_response = request_bbb_api("create", parameters)
+    api_response = request_api("create", parameters)
     if not api_response.get("message"):
         api_response["message"] = "Meeting created."
+    meeting.started = True
+    meeting.moderator_password = api_response["moderatorPW"]
+    meeting.attendee_password = api_response["attendeePW"]
+    meeting.save()
     return api_response
 
 
-def bbb_api_end(meeting, moderator=False):
-    """Call BBB API to end a meeting."""
-    password = meeting.attendee_password
-    if moderator:
-        password = meeting.moderator_password
-    parameters = {
-        "meetingID": str(meeting.meeting_id),
-        "password": password,
-    }
-    return request_bbb_api("end", parameters)
-
-
-def bbb_api_join(meeting, fullname, moderator=False):
+def join(meeting: Meeting, fullname, moderator=False):
     """Call BBB API to join a meeting."""
-    password = meeting.attendee_password
-    if moderator:
-        password = meeting.moderator_password
     parameters = {
         "fullName": fullname,
         "meetingID": str(meeting.meeting_id),
-        "password": password,
+        "password": _password(meeting, moderator),
         "redirect": "true",
     }
-    return request_bbb_api("join", parameters, prepare=True)
+    return request_api("join", parameters, prepare=True)
 
 
-def bbb_api_get_meeting_infos(meeting):
+def end(meeting: Meeting, moderator=False):
+    """Call BBB API to end a meeting."""
+    parameters = {
+        "meetingID": str(meeting.meeting_id),
+        "password": _password(meeting, moderator),
+    }
+    return request_api("end", parameters)
+
+
+def get_meeting_infos(meeting: Meeting):
     """Call BBB API to retrieve meeting informations."""
     parameters = {
         "meetingID": meeting.meeting_id,
     }
-    return request_bbb_api("getMeetingInfo", parameters)
+    try:
+        api_response = request_api("getMeetingInfo", parameters)
+        meeting.started = api_response["returncode"] == "SUCCESS"
+        meeting.save()
+        return api_response
+    except ApiMeetingException as exception:
+        meeting.started = False
+        meeting.save()
+        raise exception
